+++ changelog.md
@@ -11,8 +11,10 @@ Codion Change Log
 ### is.codion.framework.model
 - EntityEditor.nullable(attribute) removed.
 - DefaultEntityEditor bug fixed, internal mutable entity instance no longer propagated to state predicates and validators, state management refactored.
+- ComponentModels.searchModel(), EntityEditor parameter replaced with EntityConnection.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeStep.foreignKey() added, configures attributes and null inclusion.
+- SwingComponentModels.comboBox(), EntityEditor parameter replaced with EntityConnection.
 
 ## 0.18.70
 ### is.codion.common.reactive